fix: working get api

+++ Models/Cart.cs
@@ -6,7 +6,7 @@ namespace LinenManagementSystem.Models
         [Key]
         public required int CartId { get; set; }
         public required string Name { get; set; }
-        public required double Weight { get; set; }
+        public required int Weight { get; set; }
         public required string Type { get; set; }
     }
 }
